// src/generated-api/tournament/models/tournament-tournament-bracket-info-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TournamentBracketMatchDto } from './tournament-bracket-match-dto';
import { TournamentBracketParticipantDto } from './tournament-bracket-participant-dto';
import { TournamentGroupEntity } from './tournament-group-entity';
import { TournamentRoundEntity } from './tournament-round-entity';
import { TournamentStageEntity } from './tournament-stage-entity';

/**
 *
 * @export
 * @interface TournamentTournamentBracketInfoDto
 */
export interface TournamentTournamentBracketInfoDto {
  /**
   *
   * @type {Array<TournamentBracketParticipantDto>}
   * @memberof TournamentTournamentBracketInfoDto
   */
  participant: Array<TournamentBracketParticipantDto>;
  /**
   *
   * @type {Array<TournamentStageEntity>}
   * @memberof TournamentTournamentBracketInfoDto
   */
  stage: Array<TournamentStageEntity>;
  /**
   *
   * @type {Array<TournamentGroupEntity>}
   * @memberof TournamentTournamentBracketInfoDto
   */
  group: Array<TournamentGroupEntity>;
  /**
   *
   * @type {Array<TournamentRoundEntity>}
   * @memberof TournamentTournamentBracketInfoDto
   */
  round: Array<TournamentRoundEntity>;
  /**
   *
   * @type {Array<TournamentBracketMatchDto>}
   * @memberof TournamentTournamentBracketInfoDto
   */
  match: Array<TournamentBracketMatchDto>;
}